+++ src/components/modal/Modal.jsx
@@ -1,25 +1,16 @@
 "use client";
 
-import { useState } from "react";
 import Button from "@components/button/Button";
-import ArgsInputList from "@components/argsInput/ArgsInputList";
-import AddArgForm from "@components/argsInput/AddArgForm";
 
-const Modal = ({ isOpen, onClose, functionCode, onSubmit }) => {
-  const [args, setArgs] = useState([]);
-
-  const handleAddArg = (arg) => {
-    if (arg.trim() !== "") {
-      setArgs([...args, arg]);
-    }
-  };
-
-  const handleFormSubmit = (e) => {
-    e.preventDefault();
-    onSubmit(args);
-    onClose();
-  };
-
+export default function Modal({
+  isOpen,
+  onClose,
+  title,
+  children,
+  width = "w-1/2",
+  height = "h-1/2",
+  padding = "p-10",
+}) {
   if (!isOpen) return null;
 
   return (
@@ -28,7 +19,7 @@ const Modal = ({ isOpen, onClose, functionCode, onSubmit }) => {
       onClick={onClose}
     >
       <div
-        className="bg-[var(--modal-background-color)] flex flex-col justify-between w-3/6 h-3/6 p-10 border-indigo-500 rounded-lg relative"
+        className={`bg-[var(--modal-background-color)] flex flex-col justify-between ${width} ${height} ${padding} border-indigo-500 rounded-lg relative`}
         onClick={(e) => e.stopPropagation()}
       >
         <Button
@@ -36,24 +27,9 @@ const Modal = ({ isOpen, onClose, functionCode, onSubmit }) => {
           className="absolute top-2 right-2 text-white bg-transparent"
           onClick={onClose}
         />
-        <h1>테스트 케이스 입력</h1>
-        <form
-          onSubmit={handleFormSubmit}
-          className="flex flex-col justify-between h-full"
-        >
-          <AddArgForm onAddArg={handleAddArg} />
-          <ArgsInputList args={args} setArgs={setArgs} />
-          <div className="flex justify-end">
-            <Button
-              text="실행"
-              className="btn-purple-light px-4 py-2 rounded"
-              type="submit"
-            />
-          </div>
-        </form>
+        {title && <h1>{title}</h1>}
+        <div className="flex flex-col justify-between h-full">{children}</div>
       </div>
     </div>
   );
-};
-
-export default Modal;
+}
